feat: refactor chatstorage & fix relogin and webhook
commit 03ca43f4b9071d4a4998ad503f17e5c1b5f73565
Author: Aldino Kemal <aldinokemal2104@gmail.com>
Date:   Sun Jul 13 11:22:14 2025 +0700

    refactor(chatstorage): remove unused methods and simplify codebase

    - Remove unused interface methods: UpdateChatLastMessage, GetMessage, GetMediaInfo, etc.
    - Replace interface{} with any type alias throughout codebase
    - Add helper methods scanMessage, scanChat, getCount to reduce duplication
    - Update NewStorageRepository to return interface instead of concrete type
    - Remove unused whatsappDB parameter from app usecase initialization
    - Clean up WhatsApp client management and remove obsolete remote logout handling

commit 921a9a84be195c1c68906d9b67008da942ea001a
Author: Aldino Kemal <aldinokemal2104@gmail.com>
Date:   Sun Jul 13 10:52:46 2025 +0700

    refactor(chatstorage): migrate to domain-driven architecture with clean interfaces

    Restructure chat storage system to follow domain-driven design principles:
    - Move models and interfaces from infrastructure to domain layer
    - Implement IChatStorageRepository interface for better abstraction
    - Consolidate SQLite repository implementation with enhanced functionality
    - Remove redundant initialization and integration files
    - Update dependency injection across all layers to use domain interfaces
    - Simplify configuration by removing CSV storage path references
