+++ src/usecase/chat.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	domainChat "github.com/aldinokemal/go-whatsapp-web-multidevice/domains/chat"
-	"github.com/aldinokemal/go-whatsapp-web-multidevice/infrastructure/chatstorage"
+	domainChatStorage "github.com/aldinokemal/go-whatsapp-web-multidevice/domains/chatstorage"
 	"github.com/aldinokemal/go-whatsapp-web-multidevice/infrastructure/whatsapp"
 	"github.com/aldinokemal/go-whatsapp-web-multidevice/pkg/utils"
 	"github.com/aldinokemal/go-whatsapp-web-multidevice/validations"
@@ -15,10 +15,10 @@ import (
 )
 
 type serviceChat struct {
-	chatStorageRepo *chatstorage.Storage
+	chatStorageRepo domainChatStorage.IChatStorageRepository
 }
 
-func NewChatService(chatStorageRepo *chatstorage.Storage) domainChat.IChatUsecase {
+func NewChatService(chatStorageRepo domainChatStorage.IChatStorageRepository) domainChat.IChatUsecase {
 	return &serviceChat{
 		chatStorageRepo: chatStorageRepo,
 	}
@@ -30,7 +30,7 @@ func (service serviceChat) ListChats(ctx context.Context, request domainChat.Lis
 	}
 
 	// Create filter from request
-	filter := &chatstorage.ChatFilter{
+	filter := &domainChatStorage.ChatFilter{
 		Limit:      request.Limit,
 		Offset:     request.Offset,
 		SearchName: request.Search,
@@ -38,14 +38,14 @@ func (service serviceChat) ListChats(ctx context.Context, request domainChat.Lis
 	}
 
 	// Get chats from storage
-	chats, err := service.chatStorageRepo.Repository().GetChats(filter)
+	chats, err := service.chatStorageRepo.GetChats(filter)
 	if err != nil {
 		logrus.WithError(err).Error("Failed to get chats from storage")
 		return response, err
 	}
 
 	// Get total count for pagination
-	totalCount, err := service.chatStorageRepo.Repository().GetTotalChatCount()
+	totalCount, err := service.chatStorageRepo.GetTotalChatCount()
 	if err != nil {
 		logrus.WithError(err).Error("Failed to get total chat count")
 		// Continue with partial data
@@ -91,7 +91,7 @@ func (service serviceChat) GetChatMessages(ctx context.Context, request domainCh
 	}
 
 	// Get chat info first
-	chat, err := service.chatStorageRepo.Repository().GetChat(request.ChatJID)
+	chat, err := service.chatStorageRepo.GetChat(request.ChatJID)
 	if err != nil {
 		logrus.WithError(err).WithField("chat_jid", request.ChatJID).Error("Failed to get chat info")
 		return response, err
@@ -101,7 +101,7 @@ func (service serviceChat) GetChatMessages(ctx context.Context, request domainCh
 	}
 
 	// Create message filter from request
-	filter := &chatstorage.MessageFilter{
+	filter := &domainChatStorage.MessageFilter{
 		ChatJID:   request.ChatJID,
 		Limit:     request.Limit,
 		Offset:    request.Offset,
@@ -127,17 +127,17 @@ func (service serviceChat) GetChatMessages(ctx context.Context, request domainCh
 	}
 
 	// Get messages from storage
-	var messages []*chatstorage.Message
+	var messages []*domainChatStorage.Message
 	if request.Search != "" {
 		// Use search functionality if search query is provided
-		messages, err = service.chatStorageRepo.Repository().SearchMessages(request.ChatJID, request.Search, request.Limit)
+		messages, err = service.chatStorageRepo.SearchMessages(request.ChatJID, request.Search, request.Limit)
 		if err != nil {
 			logrus.WithError(err).WithField("chat_jid", request.ChatJID).Error("Failed to search messages")
 			return response, err
 		}
 	} else {
 		// Use regular filter
-		messages, err = service.chatStorageRepo.Repository().GetMessages(filter)
+		messages, err = service.chatStorageRepo.GetMessages(filter)
 		if err != nil {
 			logrus.WithError(err).WithField("chat_jid", request.ChatJID).Error("Failed to get messages")
 			return response, err
@@ -145,7 +145,7 @@ func (service serviceChat) GetChatMessages(ctx context.Context, request domainCh
 	}
 
 	// Get total message count for pagination
-	totalCount, err := service.chatStorageRepo.Repository().GetChatMessageCount(request.ChatJID)
+	totalCount, err := service.chatStorageRepo.GetChatMessageCount(request.ChatJID)
 	if err != nil {
 		logrus.WithError(err).WithField("chat_jid", request.ChatJID).Error("Failed to get message count")
 		// Continue with partial data
