+++ src/views/index.html
@@ -316,12 +316,6 @@
                         case 'LIST_DEVICES':
                             this.connected_devices = message.result
                             break;
-                        case 'REMOTE_LOGOUT':
-                            // Handle remote logout from phone
-                            showErrorInfo('⚠️ ' + message.message)
-                            this.connected_devices = null
-                            console.log('[REMOTE_LOGOUT] User logged out from phone:', message)
-                            break;
                         case 'LOGOUT_COMPLETE':
                             // Handle successful cleanup after remote logout
                             showSuccessInfo('✅ ' + message.message)
